requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-colorfield==0.1.15
django-cors-headers==2.5.2
django-debug-toolbar==1.11
django-filter==2.1.0
django-import-export==1.2.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
# fw==1.7.4
idna==2.8
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
# pyctivex==2.3.0
PyJWT==1.7.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
